requirements.txt
fastapi
uvicorn
langchain
langchain-openai
langchain-groq
langchain-community
httpx
python-dotenv
faiss-cpu
pydantic
